Refactor and document CommandsClass.java
The CommandsClass was refactored to improve readability and maintainability. Documentation comments were added to clearly explain the method's purpose and functionality. This will increase understandability of the code and ease future modifications or updates. The switch-statements were also modified to be case-insensitive for improved user experience.

Two new classes: StringJoiner and StringSimplifier, which are simply classes that allow you to join and simplify Strings. The join methods are currently only used in CommandsClass, however we can expand it to other classes later on too.